Split into Factory and package methods.
