+++ gazeplay-picto-pick/src/main/java/pictogrammes/Pictogrammes.java
@@ -7,6 +7,7 @@ import javafx.scene.paint.Color;
 import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
 import net.gazeplay.commons.gaze.SecondScreen;
+import net.gazeplay.commons.gaze.SecondScreenFactory;
 import net.gazeplay.commons.gaze.devicemanager.GazeDeviceManager;
 import net.gazeplay.commons.gaze.devicemanager.GazeDeviceManagerFactory;
 
@@ -42,6 +43,6 @@ public class Pictogrammes extends Application {
 
         primaryStage.show();
 
-        SecondScreen.launch();
+        SecondScreenFactory.launch();
     }
 }
